requirements.txt
flask
flask-cors
google-genai
supabase
python-dotenv